Convert search input to TS and provide clear (x) icon if text is present.
